finalises primary issue, that each boid holds its own copy of dbNeighbours, there needs to be a global, matrix?

+++ exercise2/boid.hpp
@@ -64,7 +64,7 @@ public:
 	inCluster = boid.inCluster;
 	dbNeighbours = boid.dbNeighbours;
 
-	for(int i = 0; i < boid.dbNeighbours.size(); i++){
+	for(size_t i = 0; i < boid.dbNeighbours.size(); i++){
 	  dbNeighbours[i] = boid.dbNeighbours[i];
 	}
